started refactoring to python standards
