+++ gw2_tracker/__init__.py
@@ -1,0 +1,8 @@
+__version__ = "0.1.0"
+
+from gw2_tracker import app
+
+
+def main():
+    instance = app.GW2Tracker()
+    instance.mainloop()
