+++ gw2_tracker/__main__.py
@@ -1,0 +1,1 @@
+
