+++ gw2_tracker/assets/copper_coin_20px.png
