+++ gw2_tracker/assets/copper_coin_highres.png
