+++ gw2_tracker/assets/gold_coin_highres.png
