+++ gw2_tracker/assets/silver_coin_20px.png
