+++ gw2_tracker/assets/silver_coin_highres.png
