+++ pyproject.toml
@@ -1,0 +1,28 @@
+[tool.poetry]
+name = "gw2_tracker"
+version = "0.1.0"
+description = ""
+authors = ["Krashnark"]
+include = ["assets/"]
+
+[tool.poetry.dependencies]
+python = "^3.10"
+requests = "^2.26.0"
+hypothesis = "^6.24.3"
+Pillow = "^8.4.0"
+
+[tool.poetry.dev-dependencies]
+pytest = "^6.2.5"
+black = "^21.10b0"
+isort = "^5.10.1"
+
+[tool.poetry.scripts]
+gw2_tracker = 'gw2_tracker:main'
+
+[tool.isort]
+profile = "black"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
+
