+++ tests/test_gw2_tracker.py
@@ -1,0 +1,5 @@
+from gw2_tracker import __version__
+
+
+def test_version():
+    assert __version__ == '0.1.0'
